backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
